tests/wpt/web-platform-tests/bluetooth/idl/idl-NavigatorBluetooth.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_false('bluetooth' in navigator);
}, 'navigator.bluetooth not available in insecure contexts');
</script>
